Ref #192: Move test classes into separate folder
