+++ test/karma/helpers/ChildClickView.js
