+++ test/karma/helpers/ClickView.js
@@ -1,5 +1,5 @@
 var _ = require('underscore');
-var TorsoView = require('./../../modules/View');
+var TorsoView = require('./../../../modules/View');
 
 module.exports = TorsoView.extend({
   events: {
